+++ src/Rtp/RtpSelector.h
@@ -64,7 +64,7 @@ public:
      * @return 是否成功
      */
     bool inputRtp(const toolkit::Socket::Ptr &sock, const char *data, size_t data_len,
-                  const struct sockaddr *addr, uint32_t *dts_out = nullptr);
+                  const struct sockaddr *addr, uint64_t *dts_out = nullptr);
 
     /**
      * 获取一个rtp处理器
